requirements.txt
pygame~=2.6.1
numpy~=2.1.3
pyrr~=0.10.3
PyOpenGL~=3.1.7
pillow~=11.0.0
glfw~=2.7.0
perlin_noise~=1.5
scipy~=1.14.1